Simplify placeNewGroup; move part-number scheduling into its own module
placeNewGroup: it spliced the new id into the sibling array, filtered it back
out to renumber, then used indexOf to recover where it landed. Compute the
insertion index directly instead and renumber around it — siblings before the
slot keep their index, those at or past it shift up by one. Behavior is
unchanged (the existing tests pin it). This leaves the groupId parameter unused,
so drop it: the function opens a slot, it doesn't need to know what fills it.
Trim the doc comment to match.

Part-number scheduling (loadPartNumbers, runPartNumberJob, selectPartNumberJobs,
PartNumberJob) moves out of load-group.ts — ~140 lines, 30% of that file — and
into load-part-numbers.ts, which already owned computePartNumbers and
nextWaveSize. All part-number loading now lives in one module and load-group.ts
is back to group orchestration. The moved code is otherwise unchanged.
